rename `numeric.d.ts` to `math.d.ts`
