+++ test-d/internal/array-max.ts
@@ -1,6 +1,6 @@
 import {expectType} from 'tsd';
 import type {ArrayMax} from '../../source/internal';
-import type {PositiveInfinity} from '../../source/numeric';
+import type {PositiveInfinity} from '../../source/math';
 
 expectType<ArrayMax<[1, 2, 5, 3, 7, -9, -5, 0]>>(7);
 expectType<ArrayMax<[1, 2, 5, 3, 7, -9, -5, 0, PositiveInfinity]>>(null! as PositiveInfinity);
