+++ test-d/internal/number-absolute.ts
@@ -1,6 +1,6 @@
 import {expectType} from 'tsd';
 import type {NumberAbsolute} from '../../source/internal';
-import {type NegativeInfinity, type PositiveInfinity} from '../../source/numeric';
+import {type NegativeInfinity, type PositiveInfinity} from '../../source/math';
 
 expectType<NumberAbsolute<3>>(3);
 expectType<NumberAbsolute<-3>>(3);
